Check that file exists

+++ install.sh
@@ -55,9 +55,13 @@ sudo tar -xzf "$temp_file" -C $temp
 echo "Removing ryzenadj and portio.so from new package and backing up existing ones before installing new package"
 echo "Do not use this version of the script if you want to use the provided copies from mengmeet, as they will not be kept up-to-date by this modified installation script."
 sudo rm "${temp}/${package}/bin/ryzenadj"
-sudo mv "${plugin_dir}/${package}/bin/ryzenadj" "${temp}/${package}/bin/ryzenadj"
+if [ -e "${plugin_dir}/${package}/bin/ryzenadj" ]; then
+  sudo mv "${plugin_dir}/${package}/bin/ryzenadj" "${temp}/${package}/bin/ryzenadj"
+fi  
 sudo rm "${temp}/${package}/backend/portio.so"
-sudo mv "${plugin_dir}/${package}/backend/portio.so" "${temp}/${package}/backend/portio.so"
+if [ -e "${plugin_dir}/${package}/backend/portio.so" ]; then
+  sudo mv "${plugin_dir}/${package}/backend/portio.so" "${temp}/${package}/backend/portio.so"
+fi
 sudo rsync -av "${temp}/${package}/" $plugin_dir --delete
 
 rm "$temp_file"
